components/chore: a componentize function bases in atomic structure
