+++ src/app/auth/register/page.tsx
@@ -2,11 +2,11 @@
 
 import { useState } from 'react'
 import { useRouter } from 'next/navigation'
-import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import { Title } from '@/components/Title'
+import { Button } from "@/components/ui/atoms/button"
+import { Input } from "@/components/ui/atoms/input"
+import { Label } from "@/components/ui/atoms/label"
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/atoms/card"
+import { Title } from '@/components/ui/molecules/Title'
 
 export default function RegisterPage() {
   const [name, setName] = useState('')
