+++ src/app/page.tsx
@@ -1,21 +1,25 @@
 'use client'
 
 import { useTransactions } from '@/hooks/useTransactions'
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/atoms/card"
 import { ArrowDownIcon, ArrowUpIcon, DollarSign, LogIn, LogOut, Info } from 'lucide-react'
 import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer, PieChart, Pie, Cell, LineChart, Line, AreaChart, Area } from 'recharts'
-import { AddIncomeDialog } from '@/components/AddIncomeDialog'
-import { AddExpenseDialog } from '@/components/AddExpenseDialog'
+import { AddIncomeDialog } from '@/components/ui/organisms/AddIncomeDialog'
+import { AddExpenseDialog } from '@/components/ui/organisms/AddExpenseDialog'
 import { ITransaction } from '@/interfaces/ITransaction'
-import { SummaryCard } from '@/components/SummaryCard'
-import { TransactionsTable } from '@/components/TransactionsTable'
-import { Button } from '@/components/ui/button'
+import { SummaryCard } from '@/components/ui/molecules/SummaryCard'
+import { TransactionsTable } from '@/components/ui/molecules/TransactionsTable'
+import { Button } from '@/components/ui/atoms/button'
 import { useRouter } from 'next/navigation'
-import { Title } from '@/components/Title'
+import { Title } from '@/components/ui/molecules/Title'
 import { motion } from 'framer-motion'
 import { useState } from 'react'
-import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
+import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/atoms/popover'
 import { ExpensePrediction } from '@/components/ai/ExpensePrediction'
+import {CashFlowChart} from "@/components/ui/charts/CashFlowChart";
+import {DistributionChart} from "@/components/ui/charts/DistributionChart";
+import {RecentTransactionsChart} from "@/components/ui/charts/RecentTransactionChart";
+import {IncomeVsExpensesChart} from "@/components/ui/charts/IncomeVsExpensesChart";
 
 const COLORS = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042', '#8884D8']
 
@@ -119,238 +123,86 @@ export default function DashboardFinanceiro() {
 
       <main className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
         <motion.div
-          initial={{ opacity: 0, y: -20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5 }}
-          className="flex justify-between items-center mb-8"
+            initial={{opacity: 0, y: -20}}
+            animate={{opacity: 1, y: 0}}
+            transition={{duration: 0.5}}
+            className="flex justify-between items-center mb-8"
         >
           <div>
             <h1 className="text-3xl font-bold text-gray-900">Dashboard Financeiro</h1>
             <p className="mt-1 text-sm text-gray-600">Visão geral das suas finanças pessoais</p>
           </div>
           <div className="space-x-2">
-            <AddIncomeDialog onAddIncome={handleAddIncome} />
-            <AddExpenseDialog onAddExpense={handleAddExpense} />
+            <AddIncomeDialog onAddIncome={handleAddIncome}/>
+            <AddExpenseDialog onAddExpense={handleAddExpense}/>
             <ExpensePrediction transactions={transactions}/>
           </div>
         </motion.div>
 
         <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5, delay: 0.2 }}
-          className="grid grid-cols-1 gap-6 mb-8 sm:grid-cols-2 lg:grid-cols-3"
+            initial={{opacity: 0, y: 20}}
+            animate={{opacity: 1, y: 0}}
+            transition={{duration: 0.5, delay: 0.2}}
+            className="grid grid-cols-1 gap-6 mb-8 sm:grid-cols-2 lg:grid-cols-3"
         >
           <SummaryCard
-            title="Saldo Total"
-            value={balance}
-            icon={DollarSign}
-            description="Atualizado agora"
+              title="Saldo Total"
+              value={balance}
+              icon={DollarSign}
+              description="Atualizado agora"
           />
           <SummaryCard
-            title="Receitas"
-            value={totalIncome}
-            icon={ArrowUpIcon}
-            valueColor="text-green-600"
-            description={`+${((totalIncome / (totalIncome + totalExpense)) * 100).toFixed(1)}% do total`}
+              title="Receitas"
+              value={totalIncome}
+              icon={ArrowUpIcon}
+              valueColor="text-green-600"
+              description={`+${((totalIncome / (totalIncome + totalExpense)) * 100).toFixed(1)}% do total`}
           />
           <SummaryCard
-            title="Despesas"
-            value={totalExpense}
-            icon={ArrowDownIcon}
-            valueColor="text-red-600"
-            description={`-${((totalExpense / (totalIncome + totalExpense)) * 100).toFixed(1)}% do total`}
+              title="Despesas"
+              value={totalExpense}
+              icon={ArrowDownIcon}
+              valueColor="text-red-600"
+              description={`-${((totalExpense / (totalIncome + totalExpense)) * 100).toFixed(1)}% do total`}
           />
         </motion.div>
         <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5, delay: 1.2 }}
+            initial={{opacity: 0, y: 20}}
+            animate={{opacity: 1, y: 0}}
+            transition={{duration: 0.5, delay: 1.2}}
         >
           <Card className="bg-white shadow-lg mb-8">
+            <CardTitle className="text-lg font-semibold text-gray-900">Todas as Transações</CardTitle>
+            {/*
             <CardHeader className="flex flex-row items-center justify-between">
-              <CardTitle className="text-lg font-semibold text-gray-900">Todas as Transações</CardTitle>
+
               <div className="flex space-x-2">
-                <Button onClick={() => filterTransactions('week')} variant={selectedTimeRange === 'week' ? 'default' : 'outline'}>Semana</Button>
-                <Button onClick={() => filterTransactions('month')} variant={selectedTimeRange === 'month' ? 'default' : 'outline'}>Mês</Button>
-                <Button onClick={() => filterTransactions('year')} variant={selectedTimeRange === 'year' ? 'default' : 'outline'}>Ano</Button>
-                <Button onClick={() => filterTransactions('all')} variant={selectedTimeRange === 'all' ? 'default' : 'outline'}>Todas</Button>
+                <Button onClick={() => filterTransactions('week')}
+                        variant={selectedTimeRange === 'week' ? 'default' : 'outline'}>Semana</Button>
+                <Button onClick={() => filterTransactions('month')}
+                        variant={selectedTimeRange === 'month' ? 'default' : 'outline'}>Mês</Button>
+                <Button onClick={() => filterTransactions('year')}
+                        variant={selectedTimeRange === 'year' ? 'default' : 'outline'}>Ano</Button>
+                <Button onClick={() => filterTransactions('all')}
+                        variant={selectedTimeRange === 'all' ? 'default' : 'outline'}>Todas</Button>
               </div>
             </CardHeader>
+            */}
             <CardContent>
               <TransactionsTable
-              transactions={transactions}
-              onEditTransaction={handleEditTransaction}
-              onDeleteTransaction={handleDeleteTransaction} />
+                  transactions={transactions}
+                  onEditTransaction={handleEditTransaction}
+                  onDeleteTransaction={handleDeleteTransaction}/>
             </CardContent>
           </Card>
         </motion.div>
         <div className="grid grid-cols-1 gap-6 mb-8 lg:grid-cols-2">
-
-          <motion.div
-            initial={{ opacity: 0, scale: 0.9 }}
-            animate={{ opacity: 1, scale: 1 }}
-            transition={{ duration: 0.5, delay: 0.4 }}
-          >
-            <Card className="bg-white shadow-lg">
-              <CardHeader className="flex flex-row items-center justify-between">
-                <CardTitle className="text-lg font-semibold text-gray-900">Distribuição de Receitas e Despesas</CardTitle>
-                <Popover>
-                  <PopoverTrigger>
-                    <Info className="h-5 w-5 text-gray-500" />
-                  </PopoverTrigger>
-                  <PopoverContent>
-                    Este gráfico mostra a proporção entre suas receitas e despesas totais, permitindo uma visualização rápida do equilíbrio financeiro.
-                  </PopoverContent>
-                </Popover>
-              </CardHeader>
-              <CardContent>
-                <ResponsiveContainer width="100%" height={300}>
-                  <PieChart>
-                    <Pie
-                      data={pieChartData}
-                      cx="50%"
-                      cy="50%"
-                      labelLine={false}
-                      outerRadius={80}
-                      fill="#8884d8"
-                      dataKey="value"
-                      label={({ name, percent }) => `${name} ${(percent * 100).toFixed(0)}%`}
-                    >
-                      {pieChartData.map((entry, index) => (
-                        <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
-                      ))}
-                    </Pie>
-                    <Tooltip />
-                    <Legend />
-                  </PieChart>
-                </ResponsiveContainer>
-              </CardContent>
-            </Card>
-          </motion.div>
-          <motion.div
-            initial={{ opacity: 0, scale: 0.9 }}
-            animate={{ opacity: 1, scale: 1 }}
-            transition={{ duration: 0.5, delay: 0.6 }}
-          >
-            <Card className="bg-white shadow-lg">
-              <CardHeader className="flex flex-row items-center justify-between">
-                <CardTitle className="text-lg font-semibold text-gray-900">Últimas Transações</CardTitle>
-                <Popover>
-                  <PopoverTrigger>
-                    <Info className="h-5 w-5 text-gray-500" />
-                  </PopoverTrigger>
-                  <PopoverContent>
-                    Este gráfico de barras exibe suas transações mais recentes, permitindo uma rápida comparação entre diferentes entradas e saídas.
-                  </PopoverContent>
-                </Popover>
-              </CardHeader>
-              <CardContent>
-                <ResponsiveContainer width="100%" height={300}>
-                  <BarChart data={barChartData}>
-                    <CartesianGrid strokeDasharray="3 3" />
-                    <XAxis dataKey="name" />
-                    <YAxis />
-                    <Tooltip content={({ active, payload }) => {
-                      if (active && payload && payload.length) {
-                        const data = payload[0].payload;
-                        return (
-                          <div className="bg-white p-2 border rounded shadow">
-                            <p>{data.name}</p>
-                            <p>Valor: R$ {data.valor.toFixed(2)}</p>
-                            <p>Tipo: {data.tipo}</p>
-                            <p>Categoria: {data.tag}</p>
-                          </div>
-                        );
-                      }
-                      return null;
-                    }} />
-                    <Legend />
-                    <Bar dataKey="valor" fill="#8884d8" name="Valor" />
-                  </BarChart>
-                </ResponsiveContainer>
-              </CardContent>
-            </Card>
-          </motion.div>
+          <DistributionChart pieChartData={pieChartData} colors={COLORS}/>
+          <RecentTransactionsChart barChartData={barChartData}/>
         </div>
 
-        <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5, delay: 0.8 }}
-        >
-          <Card className="bg-white shadow-lg mb-8">
-            <CardHeader className="flex flex-row items-center justify-between">
-              <CardTitle className="text-lg font-semibold text-gray-900">Fluxo de Caixa</CardTitle>
-              <Popover>
-                <PopoverTrigger>
-                  <Info className="h-5 w-5 text-gray-500" />
-                </PopoverTrigger>
-                <PopoverContent>
-                  Este gráfico de linha mostra a evolução do seu fluxo de caixa ao longo do tempo, ajudando a identificar tendências e padrões em suas finanças.
-                </PopoverContent>
-              </Popover>
-            </CardHeader>
-            <CardContent>
-              <ResponsiveContainer width="100%" height={300}>
-                <LineChart data={lineChartData}>
-                  <CartesianGrid strokeDasharray="3 3" />
-                  <XAxis dataKey="data" />
-                  <YAxis />
-                  <Tooltip content={({ active, payload }) => {
-                    if (active && payload && payload.length) {
-                      const data = payload[0].payload;
-                      return (
-                        <div className="bg-white p-2 border rounded shadow">
-                          <p>Data: {data.data}</p>
-                          <p>Valor: R$ {Math.abs(data.valor).toFixed(2)}</p>
-                          <p>Tipo: {data.valor >= 0 ? 'Receita' : 'Despesa'}</p>
-                          <p>Categoria: {data.tag}</p>
-                        </div>
-                      );
-                    }
-                    return null;
-                  }} />
-                  <Legend />
-                  <Line type="monotone" dataKey="valor" stroke="#8884d8" name="Fluxo de Caixa" />
-                </LineChart>
-              </ResponsiveContainer>
-            </CardContent>
-          </Card>
-        </motion.div>
-
-        <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5, delay: 1 }}
-        >
-          <Card className="bg-white shadow-lg mb-8">
-            <CardHeader className="flex flex-row items-center justify-between">
-              <CardTitle className="text-lg font-semibold text-gray-900">Receitas vs Despesas</CardTitle>
-              <Popover>
-                <PopoverTrigger>
-                  <Info className="h-5 w-5 text-gray-500" />
-                </PopoverTrigger>
-                <PopoverContent>
-                  Este gráfico de área empilhada compara suas receitas e despesas ao longo do tempo, permitindo visualizar facilmente a diferença entre entradas e saídas.
-                </PopoverContent>
-              </Popover>
-            </CardHeader>
-            <CardContent>
-              <ResponsiveContainer width="100%" height={300}>
-                <AreaChart data={areaChartData}>
-                  <CartesianGrid strokeDasharray="3 3" />
-                  <XAxis dataKey="data" />
-                  <YAxis />
-                  <Tooltip />
-                  <Legend />
-                  <Area type="monotone" dataKey="receita" stackId="1" stroke="#82ca9d" fill="#82ca9d" name="Receita" />
-                  <Area type="monotone" dataKey="despesa" stackId="1" stroke="#8884d8" fill="#8884d8" name="Despesa" />
-                </AreaChart>
-              </ResponsiveContainer>
-            </CardContent>
-          </Card>
-        </motion.div>
+        <CashFlowChart lineChartData={lineChartData}/>
+        <IncomeVsExpensesChart areaChartData={areaChartData}/>
       </main>
     </div>
   )
