+++ src/components/ui/charts/DistributionChart.tsx
@@ -1,0 +1,56 @@
+import { motion } from 'framer-motion'
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/atoms/card"
+import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/atoms/popover"
+import { Info } from 'lucide-react'
+import { ResponsiveContainer, PieChart, Pie, Cell, Tooltip, Legend } from 'recharts'
+
+interface DistributionChartProps {
+    pieChartData: Array<{ name: string; value: number }>
+    colors: string[]
+}
+
+export function DistributionChart({ pieChartData, colors }: DistributionChartProps) {
+    return (
+        <motion.div
+            initial={{ opacity: 0, scale: 0.9 }}
+            animate={{ opacity: 1, scale: 1 }}
+            transition={{ duration: 0.5, delay: 0.4 }}
+        >
+            <Card className="bg-white shadow-lg">
+                <CardHeader className="flex flex-row items-center justify-between">
+                    <CardTitle className="text-lg font-semibold text-gray-900">Distribuição de Receitas e Despesas</CardTitle>
+                    <Popover>
+                        <PopoverTrigger>
+                            <Info className="h-5 w-5 text-gray-500" />
+                        </PopoverTrigger>
+                        <PopoverContent>
+                            Este gráfico mostra a proporção entre suas receitas e despesas totais, permitindo uma visualização rápida do equilíbrio financeiro.
+                        </PopoverContent>
+                    </Popover>
+                </CardHeader>
+                <CardContent>
+                    <ResponsiveContainer width="100%" height={300}>
+                        <PieChart>
+                            <Pie
+                                data={pieChartData}
+                                cx="50%"
+                                cy="50%"
+                                labelLine={false}
+                                outerRadius={80}
+                                fill="#8884d8"
+                                dataKey="value"
+                                label={({ name, percent }) => `${name} ${(percent * 100).toFixed(0)}%`}
+                            >
+                                {pieChartData.map((entry, index) => (
+                                    <Cell key={`cell-${index}`} fill={colors[index % colors.length]} />
+                                ))}
+                            </Pie>
+                            <Tooltip />
+                            <Legend />
+                        </PieChart>
+                    </ResponsiveContainer>
+                </CardContent>
+            </Card>
+        </motion.div>
+    )
+}
